Add trademark claims modal unit tests
