+++ packages/domain-search/src/hooks/use-has-scrolled-to-end.ts
@@ -1,18 +1,5 @@
 import { RefObject, useEffect, useState } from 'react';
-
-export const hasScrolledToEnd = ( {
-	scrollHeight,
-	scrollTop,
-	clientHeight,
-}: {
-	scrollHeight: number;
-	scrollTop: number;
-	clientHeight: number;
-} ) => {
-	// NOTE: scrollTop might be fractional in some browsers, so without this Math.abs() trick
-	// sometimes the result won't be a whole number, causing the comparison to fail.
-	return Math.abs( scrollHeight - clientHeight - scrollTop ) < 1;
-};
+import { hasScrolledToEnd } from '../helpers/has-scrolled-to-end';
 
 export const useHasScrolledToEnd = ( contentRef: RefObject< HTMLElement > ) => {
 	const [ hasScrolledToEndResult, setHasScrolledToEndResult ] = useState( false );
